local_requirements.txt
attrs==22.1.0
awsglue3-local==1.0.0
black==22.10.0
boto3==1.26.22
botocore==1.29.22
click==8.1.3
exceptiongroup==1.0.4
findspark==2.0.1
iniconfig==1.1.1
jmespath==1.0.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.2
platformdirs==2.5.4
pluggy==1.0.0
py4j==0.10.9.5
pyparsing==3.0.9
pyspark==3.3.1
pytest==7.2.0
pytest-mock==3.10.0
pytest-spark==0.6.0
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
tomli==2.0.1
urllib3==1.26.13
